pyproject.toml
[tool.poetry]
name = "domconnect-selenium"
version = "0.1.0"
description = "Тестовое задание Domconnect на использование Selenium"
authors = ["alexanderchainsaw <alexanderkolesov.contact@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.26.1"
pytest = "^8.3.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
